// pwa/src/generated/models/threadUpdatePayload.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wetty Chat API
 * Real-time chat application backend API supporting groups, messaging, threads, stickers, invites, and push notifications.
 * OpenAPI spec version: 0.1.0
 */
import type { SnowflakeID } from './snowflakeID';

export interface ThreadUpdatePayload {
  chatId: SnowflakeID;
  lastReplyAt: string;
  replyCount: number;
  threadRootId: SnowflakeID;
}